Ruby > Hamming: Iteration #1
I started off with some pretty concise code. It was short, but I wanted it to
be more descriptive. So I opted to go with the extracted methods and naming
that I thought was a little more explicit, telling a better story where the
person reading would hopefully be able to follow along easier (especially if
they had read the instructions).