Ruby > Hamming: Iteration #3
Took Katrina's advice and used 'take' before the 'zip'. This allows me to eliminate
some confusing methods making it definitely feel cleaner. I still don't like
something about the 'nucleotide_pairs' method, but not quite sure what it is...
maybe just the length of the line. Perhaps splitting it up with a variable
might be easier to read...