Ruby > RNA Transcription: Iteration #1
I think this is a pretty good first iteration that prevents a lot of
possible duplication that might seem difficult to avoid.

The 'conversion_table' idea came from watching Jim Weirich do the 'Roman
Numerals Kata'. It blew my mind when he was able to get rid of all those
conditionals by doing something so simple and elegant. It has always
been stuck in my head, so I'm really glad I had the chance to use it
here.

I'm not sure how others feel about using 'eval' here in the
'complement_for' method, but I figured since it wasn't direct user input
then it was probably ok. I'm interested to see if others have done it
this way as well, or if there is just a much better way to do it.
(probably!)